luci-app-autodl: bump to 2021-11-04

+++ package/network/services/luci-app-autodl/files/autodl/xmlygetlist.sh
@@ -1,9 +1,20 @@
 #!/bin/sh
 
-rm /tmp/tmp.Audioxm.list
-
 paudiopath=$(uci get autodl.@autodl[0].xmlypath)
 paudioname=$(uci get autodl.@autodl[0].xmlyname)
 
-cd /$paudiopath/$paudioname
-find *.mp3 > /tmp/tmp.Audioxm.list
+testplayer=$(opkg list-installed | grep "gst-play-1.0")
+
+cd $paudiopath/$paudioname
+
+if [ -e /tmp/tmp.Audioxm.list ];then
+	rm /tmp/tmp.Audioxm.list
+fi
+
+if [ ! "$testplayer" ];then
+	find *.mp3 > /tmp/tmp.Audioxm.list
+else
+	find *.mp3 > /tmp/tmp.Audioxm.list
+	find *.m4a >> /tmp/tmp.Audioxm.list
+	find *.aac >> /tmp/tmp.Audioxm.list
+fi
